Write to application support instead of documents dir
https://github.com/jameszaghini/floaty/issues/11

+++ floaty/Storage.swift
@@ -16,38 +16,21 @@ public class Storage {
 
     fileprivate init() { }
 
-    enum Directory {
-        case documents
-        case caches
+    static private var applicationSupportURL: URL? {
+        guard let bundleId = Bundle.main.bundleIdentifier else { return nil }
+        var applicationSupportURL = FileManager.default.urls(for: .applicationSupportDirectory, in: .userDomainMask).first
+        applicationSupportURL?.appendPathComponent(bundleId)
+        return applicationSupportURL
     }
 
-    /// Returns URL constructed from specified directory
-    static fileprivate func getURL(for directory: Directory) -> URL {
-        var searchPathDirectory: FileManager.SearchPathDirectory
-
-        switch directory {
-        case .documents:
-            searchPathDirectory = .documentDirectory
-        case .caches:
-            searchPathDirectory = .cachesDirectory
-        }
-
-        if let url = FileManager.default.urls(for: searchPathDirectory, in: .userDomainMask).first {
-            return url
-        } else {
-            fatalError("Could not create URL for specified directory!")
-        }
+    private static func createApplicationSupportDirectoryIfRequired() {
+        guard let url = Storage.applicationSupportURL else { return }
+        try? FileManager.default.createDirectory(at: url, withIntermediateDirectories: true, attributes: [:])
     }
 
-    /// Store an encodable struct to the specified directory on disk
-    ///
-    /// - Parameters:
-    ///   - object: the encodable struct to store
-    ///   - directory: where to store the struct
-    ///   - fileName: what to name the file where the struct data will be stored
-    static func store<T: Encodable>(_ object: T, to directory: Directory, as fileName: String) {
-        let url = getURL(for: directory).appendingPathComponent(fileName, isDirectory: false)
-
+    static func store<T: Encodable>(_ object: T, as fileName: String) {
+        createApplicationSupportDirectoryIfRequired()
+        guard let url = applicationSupportURL?.appendingPathComponent(fileName, isDirectory: false) else { return }
         let encoder = JSONEncoder()
         do {
             let data = try encoder.encode(object)
@@ -60,15 +43,8 @@ public class Storage {
         }
     }
 
-    /// Retrieve and convert a struct from a file on disk
-    ///
-    /// - Parameters:
-    ///   - fileName: name of the file where struct data is stored
-    ///   - directory: directory where struct data is stored
-    ///   - type: struct type (i.e. Message.self)
-    /// - Returns: Optional decoded struct model(s) of data
-    static func retrieve<T: Decodable>(_ fileName: String, from directory: Directory, as type: T.Type) -> T? {
-        let url = getURL(for: directory).appendingPathComponent(fileName, isDirectory: false)
+    static func retrieve<T: Decodable>(_ fileName: String, as type: T.Type) -> T? {
+        guard let url = applicationSupportURL?.appendingPathComponent(fileName, isDirectory: false) else { return nil }
 
         if !FileManager.default.fileExists(atPath: url.path) {
             return nil
@@ -87,9 +63,8 @@ public class Storage {
         }
     }
 
-    /// Remove specified file from specified directory
-    static func remove(_ fileName: String, from directory: Directory) {
-        let url = getURL(for: directory).appendingPathComponent(fileName, isDirectory: false)
+    static func remove(_ fileName: String) {
+        guard let url = applicationSupportURL?.appendingPathComponent(fileName, isDirectory: false) else { return }
         if FileManager.default.fileExists(atPath: url.path) {
             do {
                 try FileManager.default.removeItem(at: url)
@@ -99,9 +74,8 @@ public class Storage {
         }
     }
 
-    /// Returns BOOL indicating whether file exists at specified directory with specified file name
-    static func fileExists(_ fileName: String, in directory: Directory) -> Bool {
-        let url = getURL(for: directory).appendingPathComponent(fileName, isDirectory: false)
+    static func fileExists(_ fileName: String) -> Bool {
+        guard let url = applicationSupportURL?.appendingPathComponent(fileName, isDirectory: false) else { return false }
         return FileManager.default.fileExists(atPath: url.path)
     }
 }
